Remove an unused variable
The variable was introduced in rP9646 without being used, neither did it become used in a later commit. To me it seems that it is a left-over from a copy paste, as the ".left" functions do use the variable.
Notice rP16931 did cleanup around these lines, but did not remove this variable

Patch By: Angen
Differential Revision: https://code.wildfiregames.com/D1813



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22166 3db68df2-c116-0410-a063-a993310a9797