Move more things in the projectile part of the Attack/Ranged component
Following D945, I reckon more things should be moved from Attack/Ranged to Attack/Ranged/Projectile.

In the long run, I think most things should be moved to the Projectile, which would let us have several projectiles per attack (for example), make it easier to abstract away from the Melee/Ranged attack types, and would also be more coherent. This is the first step, as it adds the Projectile to all templates.

This also makes the launch point a parameter, though I only add "y" as implementing the rotation is left as an exercise to the reader.

Reviewed By: bb

Differential Revision: https://code.wildfiregames.com/D1171


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22184 3db68df2-c116-0410-a063-a993310a9797