Rename generic ballista template to boltshooter and onager to stonethrower to avoid ambiguity. Remove template_unit_mechanical.xml and remove '_mechanical' from ship and siege template paths.
The template tree included 'template_unit_mechanical.xml', which had two children, '*_ship.xml' and '*_siege.xml'.
This template did little and implied a relevant common ancestor to ship and siege units. As organic units do not descend from a generic 'Organic', this was not necessary.

This commit removes this template and the '_mechanical' bit from children templates to simplify.

Additionally, this renames:
 - template_unit_mechanical_siege_ballista.xml to template_unit_siege_boltshooter.xml
 - template_unit_mechanical_siege_onager.xml to template_unit_siege_stonethrower.xml
This helps avoid ambiguity, as our Roman ballista for example throws stones (even if it is a torsion engine), and 'ballista' refers to a bolt shooter in Age of Empires 1 (and in AoE 2 for the Elephant Ballista, confusingly not named a scorpion like other bolt shooters in that game).

Patch By: Nescio

Reviewers: Wraitii

Subscribers: Itms, O14 Maps, wraitii, elexis, Vulcan, Stan

Differential Revision: https://code.wildfiregames.com/D1760


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22204 3db68df2-c116-0410-a063-a993310a9797