Don't crash when calling the profiler from a 'non-main' thread
Motive behind the change: calls to Profile() currently crash unless they are triggered from the main thread, but it's somewhat difficult to know that from the code. It makes more sense to silently ignore those particularly so we can easily have the same code be threaded or not.

This also removes a few profiling calls that don't make much sense.

Differential Revision: https://code.wildfiregames.com/D1853


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22248 3db68df2-c116-0410-a063-a993310a9797