Fix UnitMotion calculation of time left to avoid units going back and forth between walking and running animations.
Because of the limited precision of our fixed-point numbers, the timeLeft calculation could sometimes return results above the actual time left, resulting in units moving a few fixed::epsilons farther than they should be, which makes them switch to the running animation. This was rather unstable however, so there was a constant 'flickering' between walking and running.
If we divide last instead of first in the operation, the errors get gobbled up by the division and we no longer have this issue.

Reported by: wowgetoffyourcellphone

Differential Revision: https://code.wildfiregames.com/D1856


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22249 3db68df2-c116-0410-a063-a993310a9797