Decouple long and hierarchical pathfinders to an extent.
Following rP22253, this decouples the hierarchical pathfinder and the long pathfinder. The long pathfinder was the class owning the hierarchical pathfinder, which didn't particularly make sense and resulted in some interface awkwardness.

At the moment, the long pathfinder still needs to hierarchical pathfinder to compute paths (to make sure they are reachable).

Differential Revision: https://code.wildfiregames.com/D1867


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22278 3db68df2-c116-0410-a063-a993310a9797