Water GLSL shader improvements around reflections and whitespace fixes.
This improves refractions around entities close to the surface, such as fishes, by handling depth better and by clipping the water plane a little lower.

This uses the skybox for reflections when refractions are enabled but reflections are disabled, making it possible to play with reflections disabled without having super-ugly water (arguably a performance improvement).

Differential Revision: https://code.wildfiregames.com/D359


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22297 3db68df2-c116-0410-a063-a993310a9797