Allow entities to be affected by Status Effects from ranged attacks.
This implements a status effects receiver component (in a similar fashion to the damage receiver component). The plan is to further extend this component notably to handle graphical indication of status effects, and a variety of other effects.

Currently implemented: ranged attacks can inflict status effects that reduce HP over time.
This can be resisted by armour.

No units currently utilise this in-game but with proper graphics support that could be changed.

Patch By: Mate-86

Reviewed By: wraitii

References #1912

Differential Revision: https://code.wildfiregames.com/D1252


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22304 3db68df2-c116-0410-a063-a993310a9797