Fix a crash on some system when Alt-tabbing during game setup.
Move checks for rendering a frame in Render() to avoid missing calls to this functions, which can crash on certain systems.

Move the sound manager's idle task out of Render().
Move the buffer swapping in Render() since we do not need to swap buffers unless we are rendering.

Patch By: Angen

Reviewed By: wraitii

Differential Revision: https://code.wildfiregames.com/D1495

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22314 3db68df2-c116-0410-a063-a993310a9797