Fix leftovers from rP22313 - improve repairing, chasing and fleeing behaviour.
rP22313 made states more caller-independent but did not change some callsites to be compatible with that. This fixes those.

This does not fix chasing entirely however since unitMotion does not always recognize that the target moved enough that the entity will never be in range.
This will be fixed upstream.

Differential Revision: https://code.wildfiregames.com/D1968


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22367 3db68df2-c116-0410-a063-a993310a9797