Ensure units do get in attacking range since that range can change over time because of elevation differences.
This is a partial revert of rP22351, which skipped the "MoveToTargetAttackingRange" in APPROACHING. I (incorrectly) assumed that the original order was still perfectly fine, but in fact the attacker's max range may have changed as that depends on the relative elevation between attack and target - and so the original order might never get us in range!
This was introduced originally in rP13626.

Add a comment to clarify this.

Further, this makes sure UnitMotion still is aware that it has a target even if it is in range from the beginning, as that could lead
to stuckiness (and did when chasing sometimes). This was done in D1984 anyways.

Fixes #5478.

Differential Revision: https://code.wildfiregames.com/D2035


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22429 3db68df2-c116-0410-a063-a993310a9797