Unit Motion - make sure units that die while moving don't keep their obstruction
Introduced by rP22355. PositionChanged messages sent between MT_Destroy handling in CmpObstruction and the actual destruction of the component could re-recreate an obstruction, which would then live forever in the obstruction manager.

This prevents that by de-activating the obstruction when it is destroyed, so that further PositionChanged messages do nothing.

Differential Revision: https://code.wildfiregames.com/D2060


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22451 3db68df2-c116-0410-a063-a993310a9797