Fix walk-to-target and other movements with max-range "0" for target entities
Entities may never get to a distance of "0" when trying to reach other entities, as that would mean they are right next to each other, or even overlapping, which can fail.
Always give some leeway to distance when trying to move to a target with no specified max-range.

This fixes WalkToTarget orders which resulted in units being stuck on mines or on dropsites occasionally.

Reported by: elexis

Fixes #5510

Differential Revision: https://code.wildfiregames.com/D2087


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22496 3db68df2-c116-0410-a063-a993310a9797