Correctly face towards target when repairing, gathering, guarding.
Calls to FaceTowardsTarget were missed in the REPAIRING, GATHERING and GUARDING unitAI state, leading occasionally to units pointing the wrong way. This was not an issue prior to rP22313, which called MoveToTargetRange, which itself called FaceTowardsTarget when it was already in range, but it now needs to be explicitly handled in unitAI.

Attacking and Healing already handled this correctly.

Fixes #5518

Differential Revision: https://code.wildfiregames.com/D2096


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22505 3db68df2-c116-0410-a063-a993310a9797