Delete unused broken JSI_IGUIObject::construct.
The constructor is invalid, because one cannot create or obtain and pass an IGUIObject C++ class instance pointer via JS, only triggering a crash when doing something with it.
If there was a use case, the GUIObject should be created with a GUIObject type specific constructor.
Remove obsolete getProperty comments.

From D1699, refs #5369 and wraitiis equivalent:
Differential Revision: https://code.wildfiregames.com/D1933



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22530 3db68df2-c116-0410-a063-a993310a9797