UnitMotion - Fix a rare pathfinding issues where units tried going straight through walls, and make sure a long path is computed even when the target is within short path or direct path range.
This fixes a regression introduced by rP22464: when an entity is ordered to move to a target within short path distance (or direct path distance), it no longer computed at least one long path, which meant it could be stuck forever if the target was not actually reachable (such as behind a wall).
To fix this, compute a long path after 3 failed computations, which should result in a delay of 1-3 turns. The previous code did this after 1 failed try - the decision to make it 3 is mostly based on the idea that in most cases, being stuck means we ran into units, not that we were ordered somewhere close. Should there be complaints, it could be lowered to 2 or 1.

This fixes a second issue, reported in #4473: units sometimes get stuck, particularly when trying to garrison a turret from 'inside' the walls. The issue is that the turret is not accessible via the inside as its obstruction + garrison range is blocked by the surrounding walls.
However, as introduced by rP17154, TryGoingStraightToTargetEntity ignores all entities with the obstruction group of the target (the reason for this being that otherwise it would never succeed, since the line towards the target would likely go through the target).
For walls and formations, this means ignoring possibly too many entities, and in the case of #4473, ignoring wall pieces. The unit thus mistakenly thought it could direct-path to the turret, and got stuck.

To fix this, we can ignore specifically the targeted entity's obstruction tag. This can be considered a fix to rP17154.

temple accepted an earlier version of this patch (specifically elexis' version).

Fixes #4473

Based on a patch by: elexis

Differential Revision: https://code.wildfiregames.com/D1424


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22533 3db68df2-c116-0410-a063-a993310a9797