Introduce CGUIColor type inheriting Color type and switch the GUI to exclusively use that.
The type differs from the Color type, because contrary to Color, it can be created from a color predefined in the GUI page (such as "yellow").
Move this predefined color check to the new class instead of hardcoding it in FromJSVal / ToJSVal, GUIUtil, JSInterface_IGUIObject and to straighten the latter.

Delete fov_wedge_color from Minimap, unused since introduction in rP1230.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22558 3db68df2-c116-0410-a063-a993310a9797