UnitMotion cleanup - remove dead code, add a common path for MoveTo functions, rename BeginPathing, move functions around for better readability.
BeginPathing renamed to ComputePathToGoal, as that is what this function does.
IsMoving renamed to IsMoveRequested, as the function returns true when the unit has a move request going on, not when it is actually moving across the map, which was misleading.

UpdateMovementState's implementation moved closer to where it is used.

PathIsShort and WAYPOINT_ADVANCE_MAX are currently unused, thus deleted.

Differential Revision: https://code.wildfiregames.com/D2067


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22568 3db68df2-c116-0410-a063-a993310a9797