Implement IGUIObject settings as a template class and IGUIObject::AddSetting as a template function.
This means the type information is available for all methods operating with the setting type, which is easier for the authors, allows for compile-time checks and optimizations.
Remove the enum that was used to indirectly obtain the setting type at runtime.
Revises SGUISetting and enum from rP290 (rP74),  std::function from rP22574.

Differential Revision: https://code.wildfiregames.com/D2145
Tested on: gcc, clang, VS2015, Jenkins



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22604 3db68df2-c116-0410-a063-a993310a9797