Prevent tree leaves from being too bright in water reflections.
Since rP18443, the sides of tree leaves are extremely bright. The reason is that they overwrite the value of the depth buffer, and thus they show the skybox in the water calculation.

The simplest fix for this is to show the skybox only at lower transparency values. A better fix would require handling alpha blending completely differently.

Refs #5125.

Differential Revision: https://code.wildfiregames.com/D2143


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22607 3db68df2-c116-0410-a063-a993310a9797