Delete wrongful proxy CGUIManager::GetPreDefinedColor from rP7214.
It is wrong because the predefined colors should be loaded from the GUI page that requests to have the color parsed, which may be different from the topmost page.
Similar to FindObjectByName removed in rP22200.

Achieve this by implementing the CGUISetting<CGUIColor>::FromJSVal specialization, moved from ScriptInterface::FromJSVal<CGUIColor>, instead of adding a CGUI pointer to CGUIColor.
Mark ScriptInterface::FromJSVal<GUIColor> and inherited CColor::ParseString explicitly as deleted, so that people get a compile error if they forget to check for predefined colors when parsing a color.

Refs #5387, D1746 > D1684 > this > r22200, rP22534, rP22558, rP22587, r22604.

Differential Revision: https://code.wildfiregames.com/D2108
Tested on: clang 8, VS2015



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22663 3db68df2-c116-0410-a063-a993310a9797