Fix OOS introduced by pathfinder threading preparation diff rP22902
Following rP22902, paths were computed at the end of turn N, and then messages were sent at the beginning of turn N+1. However, the path requests were removed at the end of turn N and so weren't serialised, and neither were computed paths. This meant rejoiners would OOS when the game was serialised with pending path results.

To fix this in preparation for threading, the architecture needs to change slightly so that requests are kept and serialised correctly, and rejoiners can compute the paths after deserialisation, in order to send the messages at the beginning of turn N+1.

Fixes #5604

Reported By: elexis

Differential Revision: https://code.wildfiregames.com/D2317


git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@22979 3db68df2-c116-0410-a063-a993310a9797