Remove virtual inheritance from the codebase, refs rP73, rP141 rP290.
Remove workaround in rP22596 / D2136, the pointer from IGUIObject to IGUITextOwner that avoided the dynamic_cast.
Have the GUIObject implementing class take responsibility of calling the virtual methods in the base classes.
Leaves the code in a clean state that can serve as an example for new GUI object classes, such as D1346, refs #4683.

Differential Revision: https://code.wildfiregames.com/D2325
Comments By: Vladislav on IRC 2019-07-30, 2019-08-22, 2019-09-20-23
Tested on: clang 8.0.1, Jenkins



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23020 3db68df2-c116-0410-a063-a993310a9797