Fix formation-walking infinite loop introduced in rP22506
rP22506 Introduced a infinite loop in which formation members keep receiving the order to formation walk to their proper position. This was due to this.formationOffset is not deleted when member gets in offset position, what means he recieved motion update and it was not because of failure.

Patch by: @Freagarach
Differential Revision: https://code.wildfiregames.com/D2424

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23217 3db68df2-c116-0410-a063-a993310a9797