Remove front step from melee attacking animations making them more natural.
The front step or the quick "close-up" the infantry does while attacking was a guide i've used while animating following the old attacking animations, however this doesn't feel natural before and neither does now. This animations follows the same attacking animation whitout getting closer to attack in order to keep the visual mesh in the root of the actor or the selection circle.

This avoid unnecesary clipping when attacking buildings.

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23355 3db68df2-c116-0410-a063-a993310a9797