[gui] building → structure
Summary:
This patch corrects occurrences of “building” into “structure”.
Whilst “constructing a building” is proper English, it is important to refer to one and the same thing with a single term, to avoid potential misunderstandings (see https://trac.wildfiregames.com/wiki/EnglishStyleGuide ), thence “building a structure”, because the unit action is “build” and the entity is a “structure” (see `simulation/templates/`).

Test Plan: Check for mistakes and omissions.

Reviewers: Gallaecio

Reviewed By: Gallaecio

Subscribers: Freagarach, O8 JS GUI

Differential Revision: https://code.wildfiregames.com/D2429

git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23366 3db68df2-c116-0410-a063-a993310a9797