Don't continue to process pending netmessages in the gamesetup_mp page after having switched to the session page already.
If there was a player rejoining, a "start" and a "playerassignments" message pending for that player, the playerassignments message would not update the assignments of the session page, but the one of the obsolete gamesetup_mp page, thus triggering "unknown player" messages for that player during the session.
This seems to only happen when two players rejoin simultaneously and with very rare timing / race condition, and thus was not reproducible without modifying the code.
Issue since rP10452.

Differential Revision: https://code.wildfiregames.com/D2559
Should fix #5660.



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@23401 3db68df2-c116-0410-a063-a993310a9797